src/batch/requirements.txt
pandas==2.0.3
numpy==1.22.0
requests== 2.22.0
fastapi==0.63.0
uvicorn==0.13.3
mysql-connector-python==8.1.0